+++ tests/test/connections_limit.rs
@@ -88,7 +88,7 @@ fn connections_limit_test(max_connections: u32)
     for client_num in 0..max_connections
     {
         // make client
-        let websocket_client = client_demo_factory().new_client(
+        let mut websocket_client = client_demo_factory().new_client(
                 client_runtime.clone(),
                 websocket_url.clone(),
                 bevy_simplenet::AuthRequest::None{ client_id: client_num as u128 },
@@ -111,7 +111,7 @@ fn connections_limit_test(max_connections: u32)
 
     // 2. connecting one more client should fail
     // make client
-    let websocket_client = client_demo_factory().new_client(
+    let mut websocket_client = client_demo_factory().new_client(
             client_runtime.clone(),
             websocket_url.clone(),
             bevy_simplenet::AuthRequest::None{ client_id: 92748u128 },
@@ -133,7 +133,7 @@ fn connections_limit_test(max_connections: u32)
     assert_eq!(websocket_server.num_connections(), max_connections as u64);
 
     // 3. disconnect one client
-    let client_to_disconnect = clients.pop().expect("there should be at least one connected client");
+    let mut client_to_disconnect = clients.pop().expect("there should be at least one connected client");
     client_to_disconnect.close();
 
     std::thread::sleep(std::time::Duration::from_millis(25));  //wait for async machinery
@@ -146,7 +146,7 @@ fn connections_limit_test(max_connections: u32)
 
     // 4. adding a client should now succeed
     // make client
-    let websocket_client = client_demo_factory().new_client(
+    let mut websocket_client = client_demo_factory().new_client(
             client_runtime.clone(),
             websocket_url.clone(),
             bevy_simplenet::AuthRequest::None{ client_id: 64819u128 },
@@ -168,7 +168,7 @@ fn connections_limit_test(max_connections: u32)
 
     // 5. connecting one more client should fail
     // make client
-    let websocket_client = client_demo_factory().new_client(
+    let mut websocket_client = client_demo_factory().new_client(
             client_runtime.clone(),
             websocket_url.clone(),
             bevy_simplenet::AuthRequest::None{ client_id: 15364898u128 },