+++ tests/test/mod.rs
@@ -5,3 +5,4 @@ mod hello_world;
 mod message_size_limit;
 mod rate_limit;
 mod request_response;
+mod synchronization;